requirements.txt
RPi.GPIO==0.7.1
sensirion-i2c-driver==1.0.0
sensirion-i2c-scd==0.1.2
spidev==3.6
